build.sh: refactor build process a bit

+++ build.sh
@@ -4,14 +4,36 @@ rm -f ratings.db
 sqlite3 ratings.db < init.sql
 
 ( cd www/ && mkdir -p content/ratings/ content/predictions/ )
-( cd www/ && rm -rf public/* )
 
-rm -f results2019.csv
+if [[ $NEWGAMES ]]; then
+  rm -f results2019.csv
+fi
 make
 
-python updateRatings.py
-python generatePages.py
+if [[ $REBUILD ]]; then
 
+( cd www/ && rm -rf content/ratings/* content/predictions/* )
+
+# 2016-17
+echo "Building 2016-17 season"
+python updateRatings.py -r results2017.csv -t teams2017.csv -n 351
+python generatePages.py -r results2017.csv
+
+# 2017-18
+echo "Building 2017-18 season"
+python updateRatings.py -r results2018.csv -t teams2018.csv -n 351
+python generatePages.py -r results2018.csv
+
+fi
+
+# 2018-19
+echo "Building 2018-19 season"
+python updateRatings.py -r results2019.csv -t teams2019.csv -n 353
+python generatePages.py -r results2019.csv
+
+( cd www/ && rm -rf public/* )
 ( cd www/ && hugo --buildFuture )
 
-s3cmd -PM sync www/public/ s3://baseline-model/
+if [[ $UPLOAD ]]; then
+  s3cmd -PM sync www/public/ s3://baseline-model/
+fi
